+++ Packages/webxr/Runtime/Scripts/WebXRControllerData.cs
@@ -20,6 +20,14 @@ namespace WebXR
     public float touchpadY;
     public float buttonA;
     public float buttonB;
+    public string[] profiles;
+  }
+
+  [System.Serializable]
+  public class WebXRControllersProfiles
+  {
+    public string[] conrtoller1;
+    public string[] conrtoller2;
   }
 
   [System.Serializable]
@@ -76,7 +84,11 @@ namespace WebXR
     public Quaternion rotation;
   }
 
-  public enum WebXRControllerHand { NONE, LEFT, RIGHT };
+  public enum WebXRControllerHand {
+    NONE = 0,
+    LEFT = 1,
+    RIGHT = 2
+  }
 
   [System.Serializable]
   public class WebXRControllerButton