+++ Packages/webxr/CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Docs page.
+- XRInputSource profiles support.
+- Added OnControllerActive, OnHandActive and OnHandUpdate actions to WebXRController.
+
+### Changed
+- WebXRController no longer instantiate hand joints.
+- Better handling for controllers buttons in the editor.
+- WebXRController TryUpdateButtons is now private.
+
+### Fixed
+- Disable controllers on XR session end.
 
 ## [0.4.1] - 2020-11-28
 ### Added